1) Created package - config; 2) Moved config files to api/config; 3) Added API versioning
