+++ api/src/main/java/naakcii/by/api/config/ApiConfigConstants.java
@@ -1,0 +1,9 @@
+package naakcii.by.api.config;
+
+public final class ApiConfigConstants {
+
+    public static final String API_V1_0 = "application/vnd.naakcii.api-v1.0+json";
+
+    private ApiConfigConstants() {
+    }
+}
