+++ api/src/main/java/naakcii/by/api/config/CorsConfig.java
@@ -1,4 +1,4 @@
-package naakcii.by.api;
+package naakcii.by.api.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
