+++ api/src/main/java/naakcii/by/api/product/controller/ProductController.java
@@ -1,9 +1,14 @@
 package naakcii.by.api.product.controller;
 
+import naakcii.by.api.config.ApiConfigConstants;
 import naakcii.by.api.product.service.ProductService;
 import naakcii.by.api.product.service.modelDTO.ProductDTO;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -15,12 +20,12 @@ public class ProductController {
     @Autowired
     ProductService productService;
 
-    @GetMapping(path = "/{id}")
+    @GetMapping(path = "/{id}", produces = ApiConfigConstants.API_V1_0)
     public List<ProductDTO> getProductsBySubcategory(@PathVariable("id") Long subcategoryId) {
         return productService.getProductsByChainIdAndSubcategoryId(subcategoryId);
     }
 
-    @GetMapping
+    @GetMapping(produces = ApiConfigConstants.API_V1_0)
     public List<ProductDTO> findProductBySubcategoryIdLazyLoading(@RequestParam("first") Integer first,
                                                                   @RequestParam("last") Integer last,
                                                                   @RequestParam("SubcategoryList") List<Long> list) {
