+++ api/src/main/java/naakcii/by/api/subcategory/controller/SubcategoryController.java
@@ -1,5 +1,6 @@
 package naakcii.by.api.subcategory.controller;
 
+import naakcii.by.api.config.ApiConfigConstants;
 import naakcii.by.api.subcategory.service.SubcategoryService;
 import naakcii.by.api.subcategory.service.modelDTO.SubcategoryDTO;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,7 +18,7 @@ public class SubcategoryController {
     @Autowired
     SubcategoryService subcategoryService;
 
-    @GetMapping(path = "/{id}")
+    @GetMapping(path = "/{id}", produces = ApiConfigConstants.API_V1_0)
     public List<SubcategoryDTO> getSubcategoriesByCategoryId(@PathVariable("id") Long id) {
         return subcategoryService.getSubcategoryByCategoryId(id);
     }
